# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/categories": {
            "get": {
                "description": "Tìm categories bằng id, nhận id dưới dạng query, trả về toàn bộ tên categories nếu để trống.",
                "tags": [
                    "Categories Controller"
                ],
                "summary": "Tìm categories bằng id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of categories, if empty then return all",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Categories"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories Controller"
                ],
                "summary": "Tạo Categories mới (Administrator only)",
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Đăng nhập bằng JSON form, trả về session token nếu thành công.",
                "tags": [
                    "Logging Controller"
                ],
                "summary": "Đăng nhập bằng JSON",
                "parameters": [
                    {
                        "description": "Login Information",
                        "name": "LoginForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginForm"
                        }
                    },
                    {
                        "description": "password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session token",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/logs": {
            "post": {
                "description": "Nhận thông tin đấu giá dưới dạng JSON, sau đó kiểm tra quyền đấu giá của user (người dùng không thể ra giá chính phiên của mình). Trả về message JSON thông báo",
                "tags": [
                    "Bidding Log Controller"
                ],
                "summary": "Ra giá trong một Session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Information to be provided",
                        "name": "NewBidInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NewBidLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/logs/:sessionid": {
            "get": {
                "description": "Nhận thông tin session_id dưới dạng Path, sau đó trả về toàn bộ lịch sử đấu giá của id đó.",
                "tags": [
                    "Bidding Log Controller"
                ],
                "summary": "Lấy lịch sử đấu giá của một Session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BidSessionLog"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/logs/last": {
            "delete": {
                "description": "Xóa lần ra giá gần nhất của phiên đấu giá.",
                "tags": [
                    "Bidding Log Controller"
                ],
                "summary": "Xóa lần ra giá gần nhất (Administrator only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Bid Log info",
                        "name": "LogInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Deletelastlog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
        "/signup": {
            "post": {
                "description": "Tạo tài khoản mới bằng JSON form, trả về session token nếu thành công.",
                "tags": [
                    "Logging Controller"
                ],
                "summary": "Tạo tài khoản mới",
                "parameters": [
                    {
                        "description": "Resgister Information",
                        "name": "RegisterInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session token",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.BidSessionLog": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createAt": {
                    "type": "string"
                },
                "sessionid": {
                    "type": "integer"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "model.Categories": {
            "type": "object",
            "properties": {
                "categoriesid": {
                    "type": "integer"
                },
                "categoriesname": {
                    "type": "string"
                }
            }
        },
        "model.Deletelastlog": {
            "type": "object",
            "required": [
                "sessionid",
                "userid"
            ],
            "properties": {
                "sessionid": {
                    "type": "integer"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "model.LoginForm": {
            "type": "object",
            "required": [
                "password",
                "userid"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "model.NewBidLog": {
            "type": "object",
            "required": [
                "amount",
                "sessionid"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "sessionid": {
                    "type": "integer"
                }
            }
        },
...
        "model.RegisterForm": {
            "type": "object",
            "required": [
                "email",
                "password",
                "userid"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
    }
}